+++ ferrodec-decimal32/src/ops/mod.rs
@@ -10,6 +10,7 @@
 
 pub(crate) mod addsub;
 pub(crate) mod div;
+pub(crate) mod fma;
 pub(crate) mod mul;
 pub(crate) mod rem;
 pub(crate) mod round;